Fixed column named bug and added  alternative hypotheses selection.
Fixed a bug where the columns (to be processed) are compared to the selected variable name using a contains instead of a compare to. Cause an error in
cases where the variable is for instance Fitness and the column name CollectiveFitness.

Added alternative hypothesis drown-down box. It is now possible to select the alternative hypothesis for a MWU test from a combo-box. Thanks
Gary.