Refactor of View class
Continuation of refactor to separate the model classes from the view. Experiments have been refactored to operate via an interface within a collection.
The collection and a number of other model related classes are encapsulated in an ExperimentAnalysisModel. The ExperimentAnalysisModel is seperated
from the View by ExperimentController and a sub-controller ExperimentTestController.

In addition a greater number of exceptions are handled and a few bugs have been addressed.